+++ geom/type_multi_polygon.go
@@ -569,3 +569,14 @@ func (m MultiPolygon) Simplify(threshold float64, nv ...NoValidate) (MultiPolygo
 	}
 	return simpl, nil
 }
+
+// Densify returns a new MultiPolygon with additional linearly interpolated
+// control points such that the distance between any two consecutive control
+// points is at most the given maxDistance.
+func (m MultiPolygon) Densify(maxDistance float64) MultiPolygon {
+	ps := make([]Polygon, len(m.polys))
+	for i, p := range m.polys {
+		ps[i] = p.Densify(maxDistance)
+	}
+	return MultiPolygon{ps, m.ctype}
+}